requirements.txt
asgiref==3.8.1
attrs==24.3.0
contourpy==1.3.1
cycler==0.12.1
deta==1.2.0
dj-database-url==2.3.0
Django==5.1.4
django-cleanup==9.0.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.2
django-cors-headers==4.6.0
drf-spectacular==0.28.0
fonttools==4.55.3
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
packaging==24.2
Panda3D==1.10.15
panda3d-gltf==1.2.1
panda3d-simplepbr==0.12.0
pillow==11.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyparsing==3.2.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.22.3
screeninfo==0.8.1
six==1.17.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
